run clumppling

This function runs the clumppling algorithm.

gt_clumppling(
  input_path,
  input_format = "admixture",
  cd_param = 1,
  use_rep = 0,
  merge_cls = 0,
  cd_default = 1,
  output_path = tempfile("clump_out")
)

Arguments

input_path

the path where the Q files are stored, either a directory or a zip archive, or a q_matrix_list object

input_format

a string defining the format of the input files, one of 'admixture' (default)

cd_param

the cd_param use_rep boolean on whether a representative repeat should be used as a consesus for a mode. Defaults to FALSE, which leads to the computatoin of an average

use_rep

boolean

merge_cls

boolean,

cd_default

boolean

output_path

(optional) the clumppling functions in python save everything to file. By default, R stores the information in objects in the environment, and sends those files to a temporary directory that will be cleared at the end of a session. output_path allows to change the location of those files. This is only useful to those interested in recovering the same files as created by the python clumppling module, or for debugging.

Value

a list of class clumpling TODO describe the elements of the list

Details

If you would like to generate an annotated autoplot from your gt_clumppling object, ensure that all individuals from the same population are adjacent to one another in the Q-matrix or gt_admix object supplied to gt_clumppling. Autoplot 'group' argument requires that all individuals from the same group are adjacent.